requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astunparse==1.6.3
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==7.1.2
cloudpickle==3.0.0
Cython==3.0.11
demucs==4.0.1
diffq==0.2.4
dora_search==0.1.12
einops==0.8.0
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==2.0.3
flatbuffers==1.12
fsspec==2024.6.1
future==1.0.0
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.66.1
h11==0.12.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
julius==0.2.7
keras==2.9.0
Keras-Preprocessing==1.1.2
lameenc==1.7.0
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
mutagen==1.47.0
networkx==3.3
norbert==0.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openunmix==1.3.0
opt-einsum==3.3.0
packaging==24.1
pandas==1.5.3
protobuf==3.19.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3986==1.5.0
rsa==4.9
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
spleeter==2.4.0
submitit==1.5.1
sympy==1.13.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
torch==2.4.0
torchaudio==2.4.0
tqdm==4.66.5
treetable==0.2.5
triton==3.0.0
typer==0.3.2
typing_extensions==4.12.2
urllib3==2.2.2
websockets==13.0.1
Werkzeug==2.0.3
wrapt==1.16.0
yt-dlp==2024.8.6
